refactor(application): move work evidence retrieval out of root
