+++ crates/tracedecay/src/daemon/invocation_tests/work_evidence_journey_tests.rs
@@ -333,18 +333,14 @@ async fn registered_work_evidence_hydrates_the_provider_qualified_task_session()
         )
         .expect("mounted session retrieval");
     let evidence_retrieval =
-        crate::daemon::work_evidence_retrieval::DaemonWorkEvidenceRetrievalV1::new(Arc::new(
-            retrieval,
-        ))
-        .with_federated_authority(Arc::new(
-            crate::daemon::work_evidence_retrieval::tests::StaticFederatedAuthority(Arc::new(
-                crate::daemon::work_evidence_retrieval::tests::federated_authority(id::<
-                    PrivacyDomainId,
-                >(
-                    "privacy.work.evidence-journey",
+        tracedecay_application::work::WorkTaskSessionEvidenceRetrievalV1::new(Arc::new(retrieval))
+            .with_federated_authority(Arc::new(
+                tracedecay_application::work::StaticFederatedAuthority(Arc::new(
+                    tracedecay_application::work::federated_authority(id::<PrivacyDomainId>(
+                        "privacy.work.evidence-journey",
+                    )),
                 )),
-            )),
-        ));
+            ));
 
     let actor = id::<ActorId>("actor.work.evidence-journey");
     let grant_digest = digest('d');
